LayoutTests/canvas/philip/tests/2d.path.bezierCurveTo.nonfinite-expected.txt

< [index] >
2d.path.bezierCurveTo.nonfinite
bezierCurveTo() with Infinity/NaN is ignored
References: 2d.nonfinite
Actual output:
Expected output:

Passed
